Add confirm dialog on delete entity action.

+++ template/core-dist/src/app/pages/abstract-entity-index.component.ts
@@ -1,9 +1,17 @@
 import {ActivatedRoute, Router} from '@angular/router';
 import {AbstractEntityService} from '../shared/services/abstract-entity.service';
+import {DeleteDialogComponent} from './delete-dialog.component';
+import {MdDialog, MdDialogRef} from '@angular/material';
 
 export abstract class AbstractEntityIndexComponent<T> {
+  private deleteDialogRef: MdDialogRef<DeleteDialogComponent>;
 
-  constructor(private service: AbstractEntityService<T>, private router: Router, private route: ActivatedRoute) {}
+  constructor(
+    private service: AbstractEntityService<T>,
+    private router: Router,
+    private route: ActivatedRoute,
+    private deleteDialog: MdDialog
+  ) {}
 
   viewEntity(id: string): void {
     this.router.navigate(['view', id], { relativeTo: this.route});
@@ -14,7 +22,14 @@ export abstract class AbstractEntityIndexComponent<T> {
   }
 
   deleteEntity(id: string): void {
-    this.service.deleteEntity(id);
+    this.deleteDialogRef = this.deleteDialog.open(DeleteDialogComponent, { disableClose : true });
+
+    this.deleteDialogRef.afterClosed().subscribe(result => {
+      this.deleteDialogRef = null;
+      if (result.success) {
+        this.service.deleteEntity(id);
+      }
+    });
   }
 
   addEntity(): void {
